adds: ability to ping candig api
removes: async stuff that does not work anyway.

It is unfortunate that I did not perform
`git fetch` soon enough and only learned that a
bug had been fixed in `candig-server` that solved
the `KeyError` for me.

This change makes possible to fetch all the
data for all the available datasets in candig-server.