Dockerfile and JWT sub based rate limiting (#7)
* adds: Sessions based tracking for requests

A user has to be logged in and has to send Authorization header
along with their request to this endpoint. Since this will live
behind Tyk which will sent the "Authorization: Bearer ..." header
to this application, this is likely ok.

Still pending:

- Actual checks of JWT (to verify and validate)
- Session oriented response, if a key has data present in the session, send that

* docs: updates CHANGELOG.md

* docs (AUTHORS.md): converts to md from rst

* fix (session): improvements based on Samarth A's suggestions

* fix (misc): removes `print()`

* adds (session): Rate limiting key func

Rate limiting using Flask-Limiter by providing JWT `sub` claim to enforce per user limiting based on the passed JWT.

* adds (misc): Dockerfile

Updates docs

* docs: checks off a TODO because we added Dockerfile

* adds: Sessions based tracking for requests

A user has to be logged in and has to send Authorization header
along with their request to this endpoint. Since this will live
behind Tyk which will sent the "Authorization: Bearer ..." header
to this application, this is likely ok.

Still pending:

- Actual checks of JWT (to verify and validate)
- Session oriented response, if a key has data present in the session, send that

* fix (session): improvements based on Samarth A's suggestions

* fix (misc): removes `print()`

* adds (session): Rate limiting key func

Rate limiting using Flask-Limiter by providing JWT `sub` claim to enforce per user limiting based on the passed JWT.

* adds (misc): Dockerfile

Updates docs

* docs: checks off a TODO because we added Dockerfile