Release 0.2.0 (#8)
* fix: docs point to correct urls

* adds: travis and tox example files

* updates: travis and shell.nix

* fix: travis config

* fix: travis config

* adds: differential privacy library as dependency

* First basic major merge

* adds: openapi

* MORE STUFF

* adds: gevent but removes async calls

* adds: ability to ping candig api

removes: async stuff that does not work anyway.

It is unfortunate that I did not perform
`git fetch` soon enough and only learned that a
bug had been fixed in `candig-server` that solved
the `KeyError` for me.

This change makes possible to fetch all the
data for all the available datasets in candig-server.

* adds: dependencies in setup.cfg

* adds: `LD_LIBRARY_PATH` update to find stdlibc++ so

* adds dpath dependency

* adds the ability to calcualate raw percentage of the data
within the given attribute of interest

* a very primitive dp sum

* wip: dp search

* wip: dp search

* removes unused imports; adds PyScaffold==3.2.3

* adds tox and basic flake8

* docs and test dependencies

* docs: README.md

* deps: dev/test

* docs: CHANGELOG.md

* adds: Tests (#3)

* docs: CHANGELOG.md

* updates (api config): closed dictionary data stores require examples

Adding examples works in the interim, as we progress towards more
robust searching capability for the user to pass any string in
`term`.

* refactor (mesi_search): a module (candig.py) was wrongly situated outside

`candig.py` was outside the `mesi_search` package which was wrong because
we need the utility to be a part of that package.

Ideally, modules like `candig.py` would be a pluggable middleware instead
of having to write our own but that optimization is NOT happening here.

* tests: candig module tests

A lot of changes in this commit are mostly refactoring
that originated out of writing unit tests for candig module.

- refactors of `candig` module to split functions to reduce their complexity
- `filter` is renamed to `data_filter` to avoid name clash with global `config`
- renames `config` to `settings` to avoid name clash
- introduces `flake8` to automatically check for python code style PEP8
- fixes PEP8 issues

* docs: CHANGELOG reflects added tests

* Shows all counts, based approach where all the data for each `attributesOfInterest` 

- reverts to counts based approach where all the data for each `attributesOfInterest` is shown to the user instead of sum/percentaged based that I was doing earlier.
- adds documentation for functions
- improves tests a little

* wip: moving to publish private data for more than one attribute of interest at a time

* fix: tests

after moving to the new diff priv method, to calculate private numbers
and serve all of the results, as per Jonathan's suggestion

* docs: removes TODO that was done

* adds: logging
fix: tests
docs: docstrings

* Count based approach 2 - tests and docs (#5)

* wip: moving to publish private data for more than one attribute of interest at a time

* fix: tests

after moving to the new diff priv method, to calculate private numbers
and serve all of the results, as per Jonathan's suggestion

* docs: removes TODO that was done

* adds: logging
fix: tests
docs: docstrings

* docs: updates CHANGELOG.md

* adds: Sessions based tracking for requests (#6)

* adds: Sessions based tracking for requests

A user has to be logged in and has to send Authorization header
along with their request to this endpoint. Since this will live
behind Tyk which will sent the "Authorization: Bearer ..." header
to this application, this is likely ok.

* Dockerfile and JWT sub based rate limiting (#7)

* adds: Sessions based tracking for requests

A user has to be logged in and has to send Authorization header
along with their request to this endpoint. Since this will live
behind Tyk which will sent the "Authorization: Bearer ..." header
to this application, this is likely ok.

Still pending:

- Actual checks of JWT (to verify and validate)
- Session oriented response, if a key has data present in the session, send that

* docs: updates CHANGELOG.md

* docs (AUTHORS.md): converts to md from rst

* fix (session): improvements based on Samarth A's suggestions

* fix (misc): removes `print()`

* adds (session): Rate limiting key func

Rate limiting using Flask-Limiter by providing JWT `sub` claim to enforce per user limiting based on the passed JWT.

* adds (misc): Dockerfile

Updates docs

* docs: checks off a TODO because we added Dockerfile

* adds: Sessions based tracking for requests

A user has to be logged in and has to send Authorization header
along with their request to this endpoint. Since this will live
behind Tyk which will sent the "Authorization: Bearer ..." header
to this application, this is likely ok.

Still pending:

- Actual checks of JWT (to verify and validate)
- Session oriented response, if a key has data present in the session, send that

* fix (session): improvements based on Samarth A's suggestions

* fix (misc): removes `print()`

* adds (session): Rate limiting key func

Rate limiting using Flask-Limiter by providing JWT `sub` claim to enforce per user limiting based on the passed JWT.

* adds (misc): Dockerfile

Updates docs

* docs: checks off a TODO because we added Dockerfile